Перенес Schema из Views в Model
