+++ config.py
@@ -2,4 +2,4 @@ class Config(object):
     SQLALCHEMY_DATABASE_URI = 'sqlite:///movies.db'
     SQLALCHEMY_TRACK_MODIFICATION = False
     DEBUG = True
-    SECRET_HERE = '249y823r9v8238r9u'
+    SECRET_KEY = '249y823r9v8238r9u'
