+++ dao/model/directors_model.py
@@ -1,7 +1,13 @@
 from setup_db import db
+from marshmallow import Schema, fields
 
 
 class Director(db.Model):
     __table_name__ = 'director'
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(255))
+
+
+class DirectorSchema(Schema):
+    id = fields.Int()
+    name = fields.Str()
