+++ implemented.py
@@ -1,4 +1,4 @@
-# файл для создания DAO и сервисов чтобы импортировать их везде
+# файл для создания DAO и сервисов, чтобы импортировать их везде
 
 # book_dao = BookDAO(db.session)
 # book_service = BookService(dao=book_dao)
